+++ tools/scripts/rockset_autoscale.py
